Revise mass set field dialog
